Lightning Staff
Added lightning staff, which will be the beholder drop.

It charges by calling lighting down upon the weilder, then can be used to fire lances of lightning at enemies, stopped only by solid blocks (or, you know, going 300 blocks or some such)

Basically the thing is hillarious with loads of charges, feasible as a weapon with a few. Smiting is gonna happen.