Beholder shenanigans
Added "/spawn", the lame way of making bosses!

Added the beholder class which is pretty neat. Handles the boss bar, the death, creation, any passive effects it might have, etc. Basically everything but rays and event reading.

Added the plague (this is maybe funny, definitely untested)

Fixed itembuilder souting lore

made the beholderlistener. That's pretty neat. When a beholder spawns it makes a beholder class for it (no check for pre-existing beholders yet... yikes). It also handles all the rays which haven't been properly tested yet. (deletion and antimagic are good, paralysis is... kinda, but also crippling so hard)

Fixed dropped items thinking they got hit whenever environmental damage happened. That was throwing errors.

various registering as well as the beholder and plague lists in main

updated plugin.yml

So... beholders currently double-ray whenever they eye ray, and single-ray whenever a player hits them. I'm fine with a thorns ability, but it needs to be changed. The double ray is a little *too* out of whack for my liking as it only targets 1 player.